Infrastructure/soil-sensor-service/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pyserial>=3.5
asyncpg>=0.29.0
redis>=5.0.0
pyyaml>=6.0

